refactor(transport): move public interfaces/exceptions to API
Moves the exception classes and the `RemoteStreamMetrics` interface to
the API package, as these are, in fact, part of the stream transport
API.
