+++ transport/src/main/java/io/camunda/zeebe/transport/stream/api/NoSuchStreamException.java
@@ -5,11 +5,6 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.transport.stream.impl;
+package io.camunda.zeebe.transport.stream.api;
 
-public class StreamExhaustedException extends RuntimeException {
-
-  public StreamExhaustedException(final String message) {
-    super(message);
-  }
-}
+public class NoSuchStreamException extends RuntimeException {}
