+++ transport/src/main/java/io/camunda/zeebe/transport/stream/api/RemoteStreamMetrics.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.transport.stream.impl;
+package io.camunda.zeebe.transport.stream.api;
 
 public interface RemoteStreamMetrics {
